cleaning up the project from Firebase db
