+++ app/src/main/java/net/squanchy/injection/ApplicationComponent.kt
@@ -8,14 +8,9 @@ import net.squanchy.remoteconfig.RemoteConfig
 import net.squanchy.remoteconfig.RemoteConfigModule
 import net.squanchy.schedule.tracksfilter.TracksFilter
 import net.squanchy.schedule.tracksfilter.TracksFilterModule
-import net.squanchy.service.firebase.FirebaseAuthService
-import net.squanchy.service.firebase.FirebaseDbService
-import net.squanchy.service.firebase.injection.FirebaseModule
+import net.squanchy.service.firestore.FirebaseAuthService
 import net.squanchy.service.firestore.FirestoreDbService
 import net.squanchy.service.firestore.injection.FirestoreModule
-import net.squanchy.service.repository.EventRepository
-import net.squanchy.service.repository.SpeakerRepository
-import net.squanchy.service.repository.TracksRepository
 import net.squanchy.service.repository.injection.RepositoryModule
 import net.squanchy.support.injection.ChecksumModule
 import net.squanchy.support.injection.CurrentTimeModule
