+++ app/src/main/java/net/squanchy/notification/NotificationService.kt
@@ -3,8 +3,7 @@ package net.squanchy.notification
 import io.reactivex.Observable
 import io.reactivex.schedulers.Schedulers
 import net.squanchy.schedule.domain.view.Event
-import net.squanchy.service.firebase.FirebaseAuthService
-import net.squanchy.service.repository.EventRepository
+import net.squanchy.service.firestore.FirebaseAuthService
 
 internal class NotificationService(private val authService: FirebaseAuthService, private val eventRepository: EventRepository) {
