+++ app/src/main/java/net/squanchy/schedule/ScheduleComponent.kt
@@ -10,7 +10,6 @@ import net.squanchy.injection.applicationComponent
 import net.squanchy.navigation.NavigationModule
 import net.squanchy.navigation.Navigator
 import net.squanchy.schedule.tracksfilter.TracksFilter
-import net.squanchy.service.repository.TracksRepository
 import net.squanchy.support.injection.CurrentTimeModule
 import net.squanchy.support.system.CurrentTime
