+++ app/src/main/java/net/squanchy/schedule/tracksfilter/TracksFilterModule.kt
@@ -3,7 +3,6 @@ package net.squanchy.schedule.tracksfilter
 import dagger.Module
 import dagger.Provides
 import net.squanchy.injection.ApplicationLifecycle
-import net.squanchy.service.repository.TracksRepository
 
 @Module
 class TracksFilterModule {
