--- app/src/main/java/net/squanchy/service/firebase/model/FirebaseEvents.kt
@@ -1,3 +1,0 @@
-package net.squanchy.service.firebase.model
-
-data class FirebaseEvents(var events: Map<String, FirebaseEvent>? = null)
