--- app/src/main/java/net/squanchy/service/firebase/model/FirebaseFavorites.kt
@@ -1,5 +1,0 @@
-package net.squanchy.service.firebase.model
-
-data class FirebaseFavorites(var favorites: Map<String, Boolean> = emptyMap()) {
-    fun hasFavorite(eventId: String) = favorites.containsKey(eventId)
-}
