--- app/src/main/java/net/squanchy/service/firebase/model/FirebaseLinks.kt
@@ -1,3 +1,0 @@
-package net.squanchy.service.firebase.model
-
-data class FirebaseLinks(var slides: String? = null)
