--- app/src/main/java/net/squanchy/service/firebase/model/FirebasePlace.kt
@@ -1,7 +1,0 @@
-package net.squanchy.service.firebase.model
-
-data class FirebasePlace(
-    var id: String? = null,
-    var name: String? = null,
-    var floor: String? = null
-)
