--- app/src/main/java/net/squanchy/service/firebase/model/FirebasePlaces.kt
@@ -1,3 +1,0 @@
-package net.squanchy.service.firebase.model
-
-class FirebasePlaces(var places: List<FirebasePlace>? = null)
