--- app/src/main/java/net/squanchy/service/firebase/model/FirebaseSpeakers.kt
@@ -1,5 +1,0 @@
-package net.squanchy.service.firebase.model
-
-class FirebaseSpeakers(
-    var speakers: List<FirebaseSpeaker>? = null
-)
