--- app/src/main/java/net/squanchy/service/firebase/model/FirebaseTrack.kt
@@ -1,9 +1,0 @@
-package net.squanchy.service.firebase.model
-
-data class FirebaseTrack(
-    var id: String? = null,
-    var name: String? = null,
-    var accent_color: String? = null,
-    var text_color: String? = null,
-    var icon_url: String? = null
-)
