--- app/src/main/java/net/squanchy/service/firebase/model/FirebaseTracks.kt
@@ -1,5 +1,0 @@
-package net.squanchy.service.firebase.model
-
-data class FirebaseTracks(
-    var tracks: List<FirebaseTrack>? = null
-)
