--- app/src/main/java/net/squanchy/service/firebase/model/FirebaseUserData.kt
@@ -1,6 +1,0 @@
-package net.squanchy.service.firebase.model
-
-data class FirebaseUserData(
-    var favorites: Map<String, Boolean>? = null,
-    var achievements: Map<String, Long>? = null
-)
