+++ app/src/main/java/net/squanchy/service/firestore/injection/FirestoreModule.kt
@@ -1,9 +1,11 @@
 package net.squanchy.service.firestore.injection
 
+import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.firestore.FirebaseFirestore
 import dagger.Module
 import dagger.Provides
 import net.squanchy.injection.ApplicationLifecycle
+import net.squanchy.service.firestore.FirebaseAuthService
 import net.squanchy.service.firestore.FirestoreDbService
 
 @Module
@@ -16,5 +18,11 @@ class FirestoreModule {
 
     @Provides
     @ApplicationLifecycle
-    internal fun firebaseDbService(database: FirebaseFirestore): FirestoreDbService = FirestoreDbService(database)
+    internal fun firestoreDbService(database: FirebaseFirestore): FirestoreDbService = FirestoreDbService(database)
+
+    @ApplicationLifecycle
+    @Provides
+    internal fun firebaseAuthService(firebaseAuth: FirebaseAuth): FirebaseAuthService = FirebaseAuthService(
+        firebaseAuth
+    )
 }
