--- app/src/main/java/net/squanchy/service/repository/TracksRepository.kt
@@ -1,9 +1,0 @@
-package net.squanchy.service.repository
-
-import io.reactivex.Observable
-import net.squanchy.schedule.domain.view.Track
-
-interface TracksRepository {
-
-    fun tracks(): Observable<List<Track>>
-}
