+++ app/src/main/java/net/squanchy/speaker/SpeakerDetailsService.kt
@@ -1,8 +1,7 @@
 package net.squanchy.speaker
 
 import io.reactivex.Observable
-import net.squanchy.service.firebase.FirebaseAuthService
-import net.squanchy.service.repository.SpeakerRepository
+import net.squanchy.service.firestore.FirebaseAuthService
 import net.squanchy.speaker.domain.view.Speaker
 
 internal class SpeakerDetailsService(
