+++ app/src/main/java/net/squanchy/venue/VenueInfoModule.kt
@@ -2,7 +2,7 @@ package net.squanchy.venue
 
 import dagger.Module
 import dagger.Provides
-import net.squanchy.service.firebase.FirebaseAuthService
+import net.squanchy.service.firestore.FirebaseAuthService
 import net.squanchy.service.firestore.FirestoreDbService
 
 @Module
