+++ app/src/test/java/net/squanchy/schedule/FirestoreScheduleServiceTest.kt
@@ -9,7 +9,7 @@ import net.squanchy.schedule.domain.view.aSchedulePage
 import net.squanchy.schedule.domain.view.aTrack
 import net.squanchy.schedule.domain.view.anEvent
 import net.squanchy.schedule.tracksfilter.TracksFilter
-import net.squanchy.service.firebase.FirebaseAuthService
+import net.squanchy.service.firestore.FirebaseAuthService
 import net.squanchy.service.firestore.FirestoreDbService
 import net.squanchy.service.firestore.aFirestoreDay
 import net.squanchy.service.firestore.aFirestoreEvent
